set up auth to uodate and delte seller

+++ craftConnect-backend/index.js
@@ -88,20 +88,34 @@ app.post("/products", verifyToken, async (req, res) => {
     }
 });
 
-app.put("/products/:id", async (req, res) => {
+app.put("/products/:id", verifyToken, async (req, res) => {
     try {
+        const role = req.user.role;
+
+        if (role !== "seller")
+            return res.status(403).json({
+                message: "Forbidden: Only sellers can update products",
+            });
+
         const { id } = req.params;
-        const { name, price, category } = req.body;
-        const product = await Product.findByIdAndUpdate(
-            id,
-            { name, price, category },
-            { returnDocument: "after" },
-        ); //new:true means updated object willl be returned
+        const product = await Product.findById(id); //new:true means updated object willl be returned
+
         if (product !== null) {
+            if (req.user.id !== product.sellerID.toString()) {
+                return res.status(403).json({
+                    message: "Forbidden: You can only delete your own products",
+                });
+            }
+            const { name, price, category } = req.body;
+            product.name = name || product.name;
+            product.price = price || product.price;
+            product.category = category || product.category;
+            await product.save();
             return res.status(200).json({
                 message: "Product updated successfully",
             });
         }
+
         return res.status(404).json({
             message: "product not found",
         });
@@ -113,16 +127,28 @@ app.put("/products/:id", async (req, res) => {
     }
 });
 
-app.delete("/products/:id", async (req, res) => {
+app.delete("/products/:id", verifyToken, async (req, res) => {
     try {
+        const role = req.user.role;
+        if (role !== "seller")
+            return res.status(403).json({
+                message: "Forbidden: Only sellers can delete products",
+            });
         const { id } = req.params;
-        const product = await Product.findByIdAndDelete(id);
+        const product = await Product.findById(id);
 
-        if (product !== null)
+        if (product !== null) {
+            if (req.user.id !== product.sellerID.toString()) {
+                return res.status(403).json({
+                    message: "Forbidden: You can only delete your own products",
+                });
+            }
+            await Product.findByIdAndDelete(id);
             return res.status(204).json({
                 //used for no content
                 message: "Product deleted successfully",
             });
+        }
         return res.status(404).json({
             message: "product not found",
         });
